Bugfix for multiple TCodes
If a report had more than one transaction and the executing user was only authorized for one of them, the class did not check if the user was actually authorized for the transaction code used to start the report.

If the transaction was started using e.g. call transaction without an explicit authority check, a user might have been able to run a transaction, that he had no S_TCODE-authorization for.

The check was fixed: if the report was started using one of its transaction codes, the authorization for that exact transaction code will be checked.